refactor(metrological_agents): refactor metrological_agents.py into a package metrological_agents and include classes into agents
