+++ agentMET4FOF/agents/__init__.py
@@ -1,2 +1,2 @@
 from .base_agents import AgentMET4FOF, AgentNetwork, MonitorAgent, AgentBuffer
-from .signal_agents import SineGeneratorAgent
+from .signal_agents import MetrologicalGeneratorAgent, SineGeneratorAgent
