+++ agentMET4FOF/agents/signal_agents.py
@@ -1,7 +1,11 @@
-from .base_agents import AgentMET4FOF
+from .base_agents import AgentMET4FOF, MetrologicalAgent
+from ..metrological_streams import (
+    MetrologicalDataStreamMET4FOF,
+    MetrologicalSineGenerator,
+)
 from ..streams.signal_streams import SineGenerator
 
-__all__ = ["SineGeneratorAgent"]
+__all__ = ["MetrologicalGeneratorAgent", "SineGeneratorAgent"]
 
 
 class SineGeneratorAgent(AgentMET4FOF):
@@ -35,3 +39,44 @@ class SineGeneratorAgent(AgentMET4FOF):
         if self.current_state == "Running":
             sine_data = self._sine_stream.next_sample()  # dictionary
             self.send_output(sine_data["quantities"])
+
+
+class MetrologicalGeneratorAgent(MetrologicalAgent):
+    """An agent streaming a specified signal
+
+    Takes samples from an instance of :py:class:`MetrologicalDataStreamMET4FOF` with sampling frequency `sfreq` and
+    signal frequency `sine_freq` and pushes them sample by sample to connected agents via its output channel.
+    """
+
+    # The datatype of the stream will be MetrologicalSineGenerator.
+    _stream: MetrologicalDataStreamMET4FOF
+
+    def init_parameters(
+        self,
+        signal: MetrologicalDataStreamMET4FOF = MetrologicalSineGenerator(),
+        **kwargs
+    ):
+        """Initialize the input data stream
+
+        Parameters
+        ----------
+        signal : MetrologicalDataStreamMET4FOF (defaults to :py:class:`MetrologicalSineGenerator`)
+            the underlying signal for the generator
+        """
+        self._stream = signal
+        super().init_parameters()
+        self.set_output_data(channel="default", metadata=self._stream.metadata)
+
+    @property
+    def device_id(self):
+        return self._stream.metadata.metadata["device_id"]
+
+    def agent_loop(self):
+        """Model the agent's behaviour
+
+        On state *Running* the agent will extract sample by sample the input
+        datastream's content and push it into its output buffer.
+        """
+        if self.current_state == "Running":
+            self.set_output_data(channel="default", data=self._stream.next_sample())
+            super().agent_loop()
